Added Main page
