+++ src/App.js
@@ -1,36 +1,19 @@
-import React, { Component, Fragment } from "react";
-import Loader from "./Components/Loader/Loader";
-import { DojotButton, CleanButton, CustomButton } from "./Components/Button";
-import { Button } from "@material-ui/core";
+import React, { Component } from "react";
 import { createMuiTheme } from "@material-ui/core/styles";
 import { ThemeProvider } from "@material-ui/styles";
-import Checkbox from "@material-ui/core/Checkbox";
-
+import CssBaseline from "@material-ui/core/CssBaseline";
+import Main from "./Pages/Main";
 import myStyles from "./theme";
 
 const miTheme = createMuiTheme(myStyles);
-
-function Checkboxes() {
-  return (
-    <Fragment>
-      <Checkbox defaultChecked />
-      <Checkbox defaultChecked color="primary" />
-      <Checkbox defaultChecked color="secondary" />
-    </Fragment>
-  );
-}
 
 class App extends Component {
   render() {
     return (
       <div className="App">
+        <CssBaseline />
         <ThemeProvider theme={miTheme}>
-          <Checkboxes />
-          <CleanButton a="x">CleanButton</CleanButton>
-          <DojotButton a="x">DojotButton</DojotButton>
-          <CustomButton a="x">CustomButton</CustomButton>
-          <Button> Material Button</Button>
-          <Loader></Loader>
+          <Main />
         </ThemeProvider>
       </div>
     );
