+++ src/theme.js
@@ -3,9 +3,15 @@
 This file is planned to contains all custom style to be applied on Material UI Theme;
 
 */
-import { red, blue } from "@material-ui/core/colors";
+import { grey, red, blue } from "@material-ui/core/colors";
 
 const myStyles = {
+  root: {
+    backgroundColor: grey["A500"]
+  },
+  grid: {
+    width: 1000
+  },
   palette: {
     primary: {
       main: blue[500]
